docs(icon): Shift location of icon (#4)
Repository level metadata should exist in .repository rather than `docs`.

Shifting the repository level metadata into .repository will make it easier for some of my portfolio and other presentation tools to collect the information it needs. Websites or docs that make use of the icon in docs/ at the moment will need to have it removed.

The rasterized icon can remain, but the source icon is removed to .repository.